requirements.txt
click
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
flask-socketio
eventlet
dnspython
gevent-websocket
psycopg2-binary